+++ modules/tolyui/CHANGELOG.md
@@ -7,5 +7,5 @@
 * [tolyui_feedback](https://pub.dev/packages/tolyui_feedback)
 * [tolyui_message](https://pub.dev/packages/tolyui_message)
 
-## 0.0.2+20
-* +widget: TolyCollapse 
+* +20: +widget: TolyCollapse 
+* +21: tolyui_navigation: 0.0.8+10